fix(ai): extract OCR into ocr-service and apply security and budget fixes
- Pull all OCR logic (worker management, image validation, PDF rendering,
  magic-byte detection) out of the 882-line rag-service.ts into a new
  ocr-service.ts; RAGService delegates to OcrService — public API unchanged
- Add header-only probeImageDimensions (PNG/JPEG/WEBP/TIFF) that enforces
  the pixel cap before loadImage decode, preventing decompression bomb OOM;
  fail fast on unparseable headers rather than decoding untrusted bytes
- Add 180s cumulative OCR wall-clock budget (OCR_TOTAL_BUDGET_MS) as the
  real cancellation bound since tesseract.js ignores AbortSignal; clamp
  each recognize call to min(45s, remaining budget)
- Reduce per-page RECOGNITION_TIMEOUT_MS from 120s to 45s so worst-case
  OCR math closes within the extraction budget
- Export OCR_MAX_IMAGE_SIZE_BYTES/MB as the canonical source of the 25MB
  limit; OCR_CONFIG.MAX_IMAGE_BYTES references it
- Add ./ocr-service subpath export to package.json
